reconfigured the setup to enable the file to run as a script.
