--- backend/mms_app_backend/api/authentication/crud.py
@@ -1,17 +1,0 @@
-from sqlalchemy.orm import Session
-from backend.mms_app_backend.api.authentication import models
-from backend.mms_app_backend.api.authentication.schemas import UserCreate
-
-
-def create_user(db: Session, user: UserCreate):
-    fake_hashed_password = user.password + "notreallyhashed"
-    db_user = models.User(email=user.email, hashed_password=fake_hashed_password)
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-
-    return db_user
-
-
-def get_user_by_email(db: Session, email: str):
-    return db.query(models.User).filter(models.User.email == email).first()
