+++ backend/mms_app_backend/mms_app_backend/api/authentication/crud.py
@@ -1,0 +1,17 @@
+from sqlalchemy.orm import Session
+from .models import User
+from .schemas import UserCreate
+
+
+def create_user(db: Session, user: UserCreate):
+    fake_hashed_password = user.password + "notreallyhashed"
+    db_user = User(email=user.email, hashed_password=fake_hashed_password)
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+
+    return db_user
+
+
+def get_user_by_email(db: Session, email: str):
+    return db.query(User).filter(User.email == email).first()
