+++ backend/mms_app_backend/mms_app_backend/api/authentication/route.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, Depends, HTTPException
 
-from backend.mms_app_backend.api.authentication.schemas import UserCreate
-from backend.mms_app_backend.api.authentication.crud import get_user_by_email, create_user
+from .schemas import UserCreate,User
+from .crud import get_user_by_email, create_user
 from sqlalchemy.orm import Session
-from backend.mms_app_backend.api.authentication.models import  Base
-from backend.mms_app_backend.database import engine
-from backend.mms_app_backend.api.utils import get_db
-from backend.mms_app_backend.api.authentication.schemas import User
+from ...database import engine,Base
+from ..utils import get_db
+
+print(__package__)
 
 engine = engine
 Base.metadata.create_all(bind=engine)
