+++ backend/mms_app_backend/mms_app_backend/database.py
@@ -12,3 +12,4 @@ engine = create_engine(
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
+
